# docs/swagger.yaml
basePath: /
definitions:
  http.apiResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/http.errorDetail'
        type: array
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
  http.createUserRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
  http.errorDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  http.healthCheckResponse:
    properties:
      message:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  http.healthResponse:
    properties:
      checked_at:
        type: string
      checks:
        items:
          $ref: '#/definitions/http.healthCheckResponse'
        type: array
      status:
        type: string
    type: object
  http.userResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Scalable Golang backend boilerplate using Gin, PostgreSQL, RabbitMQ,
    Redis, and Clean Architecture.
  title: Go Clean Architecture API
  version: "1.0"
paths:
  /health/live:
    get:
      description: Checks whether the API process is alive.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.apiResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.healthResponse'
              type: object
      summary: Liveness probe
      tags:
      - health
  /health/ready:
    get:
      description: Checks PostgreSQL, RabbitMQ, and Redis dependencies.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.apiResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.healthResponse'
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/http.apiResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.healthResponse'
              type: object
      summary: Readiness probe
      tags:
      - health
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user and publish a user.created event.
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/http.apiResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.userResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.apiResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.apiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.apiResponse'
      summary: Create user
      tags:
      - users
  /users/{id}:
    get:
      description: Retrieve a user by its numeric ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.apiResponse'
            - properties:
                data:
                  $ref: '#/definitions/http.userResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.apiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.apiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.apiResponse'
      summary: Get user by ID
      tags:
      - users
swagger: "2.0"